streamlit_app/requirements.txt
streamlit
tensorflow>=2.x
numpy
Pillow